+++ src/main/java/pl/futurecollars/invoicing/controller/invoice/InvoiceController.java
@@ -23,26 +23,26 @@ public class InvoiceController implements InvoiceApi {
     }
 
     @Override
-    public int save(@RequestBody Invoice invoice) {
+    public long save(@RequestBody Invoice invoice) {
         return invoiceService.save(invoice);
     }
 
     @Override
-    public ResponseEntity<Invoice> getById(@PathVariable int id) {
+    public ResponseEntity<Invoice> getById(@PathVariable long id) {
         return invoiceService.getById(id)
             .map(invoice -> ResponseEntity.ok().body(invoice))
             .orElse(ResponseEntity.notFound().build());
     }
 
     @Override
-    public ResponseEntity<Invoice> updateById(@PathVariable int id, @RequestBody Invoice invoice) {
+    public ResponseEntity<Invoice> updateById(@PathVariable long id, @RequestBody Invoice invoice) {
         return invoiceService.update(id, invoice)
             .map(updatedInvoice -> ResponseEntity.ok().body(updatedInvoice))
             .orElse(ResponseEntity.notFound().build());
     }
 
     @Override
-    public ResponseEntity<Invoice> deleteById(@PathVariable int id) {
+    public ResponseEntity<Invoice> deleteById(@PathVariable long id) {
         return invoiceService.delete(id)
             .map(deletedInvoice -> ResponseEntity.ok().body(deletedInvoice))
             .orElse(ResponseEntity.notFound().build());
